CloudFunction/requirements.txt
# Function dependencies, for example:
# package>=version
functions-framework==2.3.0
google-cloud-pubsub==1.7.0
google-cloud-vision==2.4.0
google-cloud-logging==2.6.0
ratelimit==2.2.1
requests-oauthlib==1.3.0
urllib3==1.26.7